requirements.txt
# Daily Chess Offline (DCO) - Python dependencies
# Python version: >=3.11,<3.13

# --- Core chess + engine integration ---
python-chess==1.999

# --- Desktop UI (Qt) ---
PySide6==6.7.3
PySide6-Addons==6.7.3
PySide6-Essentials==6.7.3

# --- Database / ORM ---
SQLAlchemy==2.0.36
alembic==1.13.3

# --- Data handling / stats ---
pandas==2.2.3
numpy==2.1.3

# --- Charts (pick one; matplotlib is simplest cross-platform) ---
matplotlib==3.9.2

# --- Utilities ---
python-dateutil==2.9.0.post0
pydantic==2.9.2

# --- Testing ---
pytest==8.3.3
pytest-cov==5.0.0

# --- Packaging (optional but recommended if you want a distributable .exe/.app) ---
pyinstaller==6.10.0

# --- Optional: prettier SVG rendering / parsing if you render python-chess SVGs ---
cairosvg==2.7.1

# --- Optional: nicer logging formatting ---
rich==13.9.4
